Update jquery_append_coaches_name.user.js
Added the ability to make coaches names a hyperlink to discord DM related to that coach. These discord links are unique to each discord user. Using this feature requires some setup but should be pretty easy. Will a how to if needed. 

Added a toggle games played button to hide games that have been played.